GS-hw: Fix Wsign-compare warning.

+++ pcsx2/GS/Renderers/HW/GSHwHack.cpp
@@ -266,7 +266,7 @@ bool GSHwHack::GSC_TombRaiderUnderWorld(GSRendererHW& r, int& skip)
 
 bool GSHwHack::GSC_BurnoutGames(GSRendererHW& r, int& skip)
 {
-	if (RFBW == 2 && std::abs(static_cast<int>(RFBP) - static_cast<int>(RZBP)) <= BLOCKS_PER_PAGE)
+	if (RFBW == 2 && std::abs(static_cast<int>(RFBP) - static_cast<int>(RZBP)) <= static_cast<int>(BLOCKS_PER_PAGE))
 	{
 		skip = 2;
 		return true;
